Split Void and Result Chain classes into separate files
Issue #25
