+++ README.md
@@ -46,8 +46,8 @@ public void Process(Report report, string author)
 	var log = new Logger();
 
 	var chain =
-		Chain
-			.CreateVoid<string, string>()
+		ChainFactory
+            .CreateVoidBuilder<string, string>()
 			.Add
 			(
 				(level, message) => level == "ERROR",
