+++ Src/ShogunLib/Patterns/ChainOfResponsibility/IVoidDo.cs
@@ -1,0 +1,104 @@
+﻿// <copyright company="XATA">
+//      Copyright (c) 2016 by Shogun, All Right Reserved
+// </copyright>
+// <author>Ivan Ivchenko</author>
+// <email>iivchenko@live.com</email>
+
+namespace ShogunLib.Patterns.ChainOfResponsibility
+{
+    /// <summary>
+    /// Defines link action with void output and no input parameters.
+    /// </summary>
+    public interface IVoidDo
+    {
+        /// <summary>
+        /// Link action.
+        /// </summary>
+        void Do();
+    }
+
+    /// <summary>
+    /// Defines link action with void output and one input parameter.
+    /// </summary>
+    /// <typeparam name="T1">Type of link first input parameter.</typeparam>
+    public interface IVoidDo<in T1>
+    {
+        /// <summary>
+        /// Link action.
+        /// </summary>
+        /// <param name="t1">First input parameter of a link.</param>
+        void Do(T1 t1);
+    }
+
+    /// <summary>
+    /// Defines link action with void output and two input parameters.
+    /// </summary>
+    /// <typeparam name="T1">Type of link first input parameter.</typeparam>
+    /// <typeparam name="T2">Type of link second input parameter.</typeparam>
+    public interface IVoidDo<in T1, in T2>
+    {
+        /// <summary>
+        /// Link action.
+        /// </summary>
+        /// <param name="t1">First input parameter of a link.</param>
+        /// <param name="t2">Second input parameter of a link.</param>
+        void Do(T1 t1, T2 t2);
+    }
+
+    /// <summary>
+    /// Defines link action with void output and three input parameters.
+    /// </summary>
+    /// <typeparam name="T1">Type of link first input parameter.</typeparam>
+    /// <typeparam name="T2">Type of link second input parameter.</typeparam>
+    /// <typeparam name="T3">Type of link third input parameter.</typeparam>
+    public interface IVoidDo<in T1, in T2, in T3>
+    {
+        /// <summary>
+        /// Link action.
+        /// </summary>
+        /// <param name="t1">First input parameter of a link.</param>
+        /// <param name="t2">Second input parameter of a link.</param>
+        /// <param name="t3">Third input parameter of a link.</param>
+        void Do(T1 t1, T2 t2, T3 t3);
+    }
+
+    /// <summary>
+    /// Defines link action with void output and four input parameters.
+    /// </summary>
+    /// <typeparam name="T1">Type of link first input parameter.</typeparam>
+    /// <typeparam name="T2">Type of link second input parameter.</typeparam>
+    /// <typeparam name="T3">Type of link third input parameter.</typeparam>
+    /// <typeparam name="T4">Type of link fourth input parameter.</typeparam>
+    public interface IVoidDo<in T1, in T2, in T3, in T4>
+    {
+        /// <summary>
+        /// Link action.
+        /// </summary>
+        /// <param name="t1">First input parameter of a link.</param>
+        /// <param name="t2">Second input parameter of a link.</param>
+        /// <param name="t3">Third input parameter of a link.</param>
+        /// <param name="t4">Fourth input parameter of a link.</param>
+        void Do(T1 t1, T2 t2, T3 t3, T4 t4);
+    }
+
+    /// <summary>
+    /// Defines link action with void output and five input parameters.
+    /// </summary>
+    /// <typeparam name="T1">Type of link first input parameter.</typeparam>
+    /// <typeparam name="T2">Type of link second input parameter.</typeparam>
+    /// <typeparam name="T3">Type of link third input parameter.</typeparam>
+    /// <typeparam name="T4">Type of link fourth input parameter.</typeparam>
+    /// <typeparam name="T5">Type of link fifth input parameter.</typeparam>
+    public interface IVoidDo<in T1, in T2, in T3, in T4, in T5>
+    {
+        /// <summary>
+        /// Link action.
+        /// </summary>
+        /// <param name="t1">First input parameter of a link.</param>
+        /// <param name="t2">Second input parameter of a link.</param>
+        /// <param name="t3">Third input parameter of a link.</param>
+        /// <param name="t4">Fourth input parameter of a link.</param>
+        /// <param name="t5">Fifth input parameter of a link.</param>
+        void Do(T1 t1, T2 t2, T3 t3, T4 t4, T5 t5);
+    }
+}
