+++ Src/ShogunLib/ShogunLib.csproj
@@ -48,15 +48,24 @@
     <Compile Include="Monads\Maybe.cs" />
     <Compile Include="Monads\MaybeExtensions.cs" />
     <Compile Include="Patterns\ChainOfResponsibility\Can.cs" />
-    <Compile Include="Patterns\ChainOfResponsibility\ChainBuilder.cs" />
-    <Compile Include="Patterns\ChainOfResponsibility\Chain.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\ChainFactory.cs" />
     <Compile Include="Patterns\ChainOfResponsibility\ICan.cs" />
-    <Compile Include="Patterns\ChainOfResponsibility\IChainBuilder.cs" />
-    <Compile Include="Patterns\ChainOfResponsibility\IDo.cs" />
-    <Compile Include="Patterns\ChainOfResponsibility\IChain.cs" />
-    <Compile Include="Patterns\ChainOfResponsibility\ILink.cs" />
-    <Compile Include="Patterns\ChainOfResponsibility\Do.cs" />
-    <Compile Include="Patterns\ChainOfResponsibility\Link.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\IResultChain.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\IResultChainBuilder.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\IResultDo.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\IResultLink.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\IVoidChain.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\IVoidChainBuilder.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\IVoidDo.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\IVoidLink.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\ResultChain.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\ResultChainBuilder.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\ResultDo.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\ResultLink.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\VoidChain.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\VoidChainBuilder.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\VoidDo.cs" />
+    <Compile Include="Patterns\ChainOfResponsibility\VoidLink.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Unit.cs" />
     <Compile Include="Units.cs" />
